Split test_reduction_ops.py into per-group files, one device per group
Rather than putting all 4,958 cases on a single module-scoped device
(6.6x the largest in-tree precedent), split the file into four groups,
each with its own module device:

  test_generic_ops.py                (2,016)  corner/dim grid
  test_generic_ops_w_scalar.py       (2,592)  scalar= grid
  test_generic_ops_memory_configs.py   (252)  ND/HEIGHT/WIDTH/BLOCK shard
                                              + dtype/layout contracts
  test_reduction_op_corners.py          (98)  topk/argmax/cumsum/cumprod/
                                              moe/sampling corners + the
                                              preallocated-run helpers

Groups are failure-isolation domains: a wedged device takes down one
group, not the whole 4,958-case collection, and the largest group is
within existing nightly precedent. The device-cycle savings are
unchanged (~5,090 of 5,096 per-case open/close cycles still
eliminated; 4 extra suite opens cost seconds).

Collection is unchanged: 4,958 cases with identical per-function
counts; directory total 5,772. Verified on BH p100a: full directory
3,978 passed / 1,794 skipped / 0 failed in 38:35.

The grouping also pre-stages the planned grid restructuring: the
(op, correction) axis pairing lands in the two generic files, and the
corner-parity rows have a dedicated destination file.
